+++ docs/.vuepress/config.js
@@ -81,7 +81,8 @@ const article = [
 ];
 
 const verisonInfo = [
-  '0.3.0'
+  '0.3.0',
+  '0.4.0'
 ];
 
 module.exports = {