+++ packages/core/src/model/node/EllipseNodeModel.ts
@@ -24,7 +24,7 @@ class EllipseNodeModel extends BaseNodeModel {
     const {
       anchorsOffset, x, y, rx, ry,
     } = this;
-    if (Array.isArray(anchorsOffset) && anchorsOffset.length > 0) {
+    if (anchorsOffset && anchorsOffset.length > 0) {
       return this.getAnchorsByOffset();
     }
     return [