+++ packages/core/src/model/node/PolygonNodeModel.ts
@@ -57,7 +57,7 @@ class PolygonNodeModel extends BaseNodeModel {
     const {
       anchorsOffset, x, y, width, height, points,
     } = this;
-    if (Array.isArray(anchorsOffset) && anchorsOffset.length > 0) {
+    if (anchorsOffset && anchorsOffset.length > 0) {
       return this.getAnchorsByOffset();
     }
     return points.map(([x1, y1]) => ({