+++ packages/core/src/model/node/RectNodeModel.ts
@@ -17,7 +17,7 @@ class RectNodeModel extends BaseNodeModel {
     const {
       anchorsOffset, x, y, width, height,
     } = this;
-    if (Array.isArray(anchorsOffset) && anchorsOffset.length > 0) {
+    if (anchorsOffset && anchorsOffset.length > 0) {
       return this.getAnchorsByOffset();
     }
     return [