+++ packages/extension/src/bpmn/events/EndEvent.ts
@@ -8,20 +8,15 @@ class EndEventModel extends CircleNodeModel {
       data.id = `Event_${getBpmnId()}`;
     }
     if (!data.text) {
-      data.text = {
-        value: '',
-        x: data.x,
-        y: data.y,
-      };
+      data.text = '';
     }
     if (data.text && typeof data.text === 'string') {
       data.text = {
         value: data.text,
         x: data.x,
-        y: data.y,
+        y: data.y + 40,
       };
     }
-    data.text.y += 40;
     super(data, graphModel);
   }
   getConnectedSourceRules() {
@@ -36,6 +31,11 @@ class EndEventModel extends CircleNodeModel {
 }
 
 class EndEventView extends CircleNode {
+  getAnchorStyle() {
+    return {
+      visibility: 'hidden',
+    };
+  }
   getShape() {
     const { x, y, fill, stroke, strokeWidth, r } = this.getAttributes();
     const outCircle = super.getShape();