+++ packages/extension/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.3](https://github.com/towersxu/logicflow/compare/@logicflow/extension@0.4.2...@logicflow/extension@0.4.3) (2021-05-16)
+
+
+### Bug Fixes
+
+* polygon case error ([6c6e821](https://github.com/towersxu/logicflow/commit/6c6e821fb5fbc2d678235264b8caca42c97f44d2))
+* remove unnecessary console ([102bd81](https://github.com/towersxu/logicflow/commit/102bd8179d5dd5f84e677ad39b209f49d2ee3a1b))
+* typo ([5326634](https://github.com/towersxu/logicflow/commit/5326634b74768b6d630e83279983a68d1265fa83))
+
+
+### Features
+
+* **extension:** endEvent node hide anchor ([6e2333f](https://github.com/towersxu/logicflow/commit/6e2333f2af8a214530a2c2ed3c1dc5597fcd68f1))
+
+
+
+
+
 ## [0.4.2](https://github.com/towersxu/logicflow/compare/@logicflow/extension@0.3.6...@logicflow/extension@0.4.2) (2021-05-13)